LayoutTests/platform/ios-sim/accessibility/internal-link.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
<script>
var successfullyParsed = false;
if (window.testRunner)
   testRunner.dumpAsText();
</script>
</head>
<body>

<a href="#name1" id="name">Name</a>

<BR><BR><BR><BR>
<BR><BR><BR>
<a name="name1">Linked Element</a>


<p id="description"></p>
<div id="console"></div>

<script>

    description("This test thats internal links return their linked elements");

    if (window.accessibilityController) {

        document.getElementById("name").focus();
        var link = accessibilityController.focusedElement;
        shouldBe("link.linkedElement().iphoneLabel", "'AXLabel: Linked Element'");
    }

    successfullyParsed = true;
</script>

<script src="../../../resources/js-test-post.js"></script>

</body>
</html>
